refactor(elements): moved toggle and toggle-group to ui/elements
